+++ util/make_cities.py
@@ -2,28 +2,29 @@
 https://geodata.lib.berkeley.edu/catalog/stanford-bx729wr3020
 """
 import datetime
-import pandas as pd
-from pandas.io.sql import read_sql
+from geopandas import read_postgis, read_parquet
 from pyiem.util import get_dbconn
+
+FILENAME = "../src/pyiem/data/geodf/cities.parquet"
 
 
 def main():
     """Go Main Go"""
     pgconn = get_dbconn("postgis")
-    df = read_sql(
+    df = read_postgis(
         """
-        SELECT st_x(geom) as lon, st_y(geom) as lat,
-        name, pop_2010 from citiesx010g WHERE pop_2010 > 500
+        SELECT geom, name, pop_2010 from citiesx010g WHERE pop_2010 > 500
         ORDER by pop_2010 DESC
         """,
         pgconn,
         index_col=None,
+        geom_col="geom",
     )
-    df.to_pickle("../src/pyiem/data/pd_cities.pickle")
+    df.to_parquet(FILENAME)
     print(df.head(20))
 
     sts = datetime.datetime.now()
-    pd.read_pickle("../src/pyiem/data/pd_cities.pickle")
+    read_parquet(FILENAME)
     ets = datetime.datetime.now()
     print((ets - sts).total_seconds())