+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this library are documented in this file.
 - Update `MapPlot.drawcities()` to no longer use `minarea` as a filter on
 which cities are potentially labeled on the map.  The bundled dataset was
 updated on use a different [upstream source](https://geodata.lib.berkeley.edu/catalog/stanford-bx729wr3020).
+- Remove internal loading support of pickled pandas DataFrames. Twas bad idea.
 
 ### New Features